+++ ui.R
@@ -19,7 +19,10 @@ QueryData<-source("external/QueryData.R",local=T)$value
 CheckData<-source("external/CheckData.R",local=T)$value
 Help<-source("external/Help.R",local=T)$value
 ViewData <- source("external/ViewData.R", local = TRUE)$value
-QAQC <- source("external/QAQC.R", local=TRUE)$value
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Tt Mod, QAQC tab, source ####
+#QAQC <- source("external/QAQC.R", local=TRUE)$value
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 shinyUI(navbarPage("WQP STORET Data Discovery Tool (QAQC Mod)", 
                    theme = "bootstrap.css",
@@ -27,7 +30,10 @@ shinyUI(navbarPage("WQP STORET Data Discovery Tool (QAQC Mod)",
                    QueryData(),
                    CheckData(),
                    ViewData(),
-                  # QAQC(),
+                   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                   # Tt Mod, QAQC tab, add ####
+                   # QAQC(),
+                   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                    Help()
                            
 ))